+++ entrypoint.sh
@@ -55,7 +55,7 @@ readonly MOUNT_PATH_RPC_PIPEFS='/var/lib/nfs/rpc_pipefs'
 
 
 ######################################################################################
-### general purpose utilities
+### logging
 ######################################################################################
 
 log() {
@@ -63,35 +63,63 @@ log() {
   echo "----> $1"
 }
 
+log_warning() {
+
+  log "WARNING: $1"
+}
+
+log_error() {
+
+  log ''
+  log "ERROR: $1"
+  log ''
+}
+
 log_header() {
 
-  echo ''
-  echo '=================================================================='
-  echo "      $1" | awk '{print toupper($0)}'
-  echo '=================================================================='
+  echo "\
+
+==================================================================
+      $(echo "$1" | awk '{print toupper($0)}')
+=================================================================="
 }
+
+
+######################################################################################
+### error handling
+######################################################################################
 
 bail() {
 
-  log "ERROR: $1"
+  log_error "$1"
   exit 1
 }
 
-warn_on_failure() {
+on_failure() {
 
   # shellcheck disable=SC2181
-  if [[ $? -ne 0 ]]; then
-    log "WARNING: $1"
+  if [[ $? -eq 0 ]]; then
+    return
   fi
+
+  case "$1" in
+    warn)
+      log_warning "$2"
+      ;;
+    stop)
+      log_error "$2"
+      stop
+      ;;
+    *)
+      bail "$2"
+      ;;
+  esac
 }
 
-exit_on_failure() {
 
-  # shellcheck disable=SC2181
-  if [[ $? -ne 0 ]]; then
-    bail "$1"
-  fi
-}
+######################################################################################
+### process control
+######################################################################################
 
 kill_process_if_running() {
 
@@ -101,7 +129,7 @@ kill_process_if_running() {
   if [[ -n $pid ]]; then
     log "killing $base"
     kill -TERM "$pid"
-    warn_on_failure "unable to kill $base"
+    on_failure warn "unable to kill $base"
   else
     log "$base was not running"
   fi
@@ -118,11 +146,11 @@ stop_mount() {
   local -r type=$(basename "$path")
 
   if mount | grep -Eq ^"$type on $path\\s+"; then
-    log "un-mounting $type from $path"
+    log "un-mounting $type filesystem from $path"
     umount -v "$path"
-    warn_on_failure "unable to un-mount $type from $path"
+    on_failure warn "unable to un-mount $type filesystem from $path"
   else
-    log "$type was not mounted on $path"
+    log "no active mount at $path"
   fi
 }
 
@@ -130,14 +158,14 @@ stop_nfsd() {
 
   log 'stopping nfsd'
   $PATH_BIN_NFSD 0
-  warn_on_failure 'unable to stop nfsd. if it had started already, check Docker host for lingering [nfsd] processes'
+  on_failure warn 'unable to stop nfsd. if it had started already, check Docker host for lingering [nfsd] processes'
 }
 
 stop_exportfs() {
 
-  log 'un-exporting filesystems'
+  log 'un-exporting filesystem(s)'
   $PATH_BIN_EXPORTFS -ua
-  warn_on_failure 'unable to un-export filesystems'
+  on_failure warn 'unable to un-export filesystem(s)'
 }
 
 stop() {
@@ -157,15 +185,6 @@ stop() {
   log_header 'terminated'
 
   exit 0
-}
-
-stop_on_failure() {
-
-  # shellcheck disable=SC2181
-  if [[ $? -ne 0 ]]; then
-    log "$1"
-    stop
-  fi
 }
 
 
@@ -244,7 +263,7 @@ assert_kernel_mod() {
 
   lsmod | grep -Eq "^$moduleName\\s+" || [ -d "/sys/module/$moduleName" ]
 
-  exit_on_failure "$moduleName module is not loaded on the Docker host's kernel (try: modprobe $moduleName)"
+  on_failure bail "$moduleName module is not loaded in the Docker host's kernel (try: modprobe $moduleName)"
 }
 
 assert_port() {
@@ -253,14 +272,14 @@ assert_port() {
   local -r value=${!envName}
 
   if [[ -n "$value" && ( "$value" -lt 1 || "$value" -gt 65535 ) ]]; then
-    bail "please set $1 to a value between 1 and 65535 inclusive"
+    bail "please set $1 to an integer between 1 and 65535 inclusive"
   fi
 }
 
 assert_nfs_version() {
 
   get_reqd_nfs_version | grep -Eq '^(3|4|4\.1|4\.2)$'
-  exit_on_failure "please set $ENV_VAR_NFS_VERSION to 3, 4, 4.1, or 4.2"
+  on_failure bail "please set $ENV_VAR_NFS_VERSION to one of: 4.2, 4.1, 4, 3"
 }
 
 assert_disabled_nfs3() {
@@ -275,7 +294,7 @@ assert_nfsd_threads() {
   local -r requested=$(get_reqd_nfsd_threads)
 
   if [[ "$requested" -lt 1 ]]; then
-    bail "please set $ENV_VAR_NFS_SERVER_THREAD_COUNT to a positive value"
+    bail "please set $ENV_VAR_NFS_SERVER_THREAD_COUNT to a positive integer"
   fi
 }
 
@@ -320,9 +339,13 @@ init_exports() {
   local candidateExportVariables
 
   candidateExportVariables=$(compgen -A variable | grep -E 'NFS_EXPORT_[0-9]+' | sort)
-  exit_on_failure "please provide $PATH_FILE_ETC_EXPORTS or set NFS_EXPORT_* environment variables"
+  on_failure bail 'failed to detect NFS_EXPORT_* variables'
 
-  log "building $PATH_FILE_ETC_EXPORTS"
+  if [[ -z "$candidateExportVariables" ]]; then
+    bail "please provide $PATH_FILE_ETC_EXPORTS to the container or set at least one NFS_EXPORT_* environment variable"
+  fi
+
+  log "building $PATH_FILE_ETC_EXPORTS from environment variables"
 
   for exportVariable in $candidateExportVariables; do
 
@@ -332,7 +355,7 @@ init_exports() {
     local dir="${lineAsArray[0]}"
 
     if [[ ! -d "$dir" ]]; then
-      log "skipping $line since $dir is not a directory"
+      log_warning "skipping $exportVariable since $dir is not a container directory"
       continue
     fi
 
@@ -348,13 +371,14 @@ init_exports() {
 
   done
 
+  log "collected $collected valid export(s) from NFS_EXPORT_* environment variables"
+
   if [[ $collected -eq 0 ]]; then
     bail 'no valid exports'
   fi
 
-  log "will export $collected filesystem(s)"
-
   echo "$exports" > $PATH_FILE_ETC_EXPORTS
+  on_failure bail "unable to write to $PATH_FILE_ETC_EXPORTS"
 }
 
 init_assertions() {
@@ -374,11 +398,11 @@ init_assertions() {
 
   # ensure /etc/exports has at least one line
   grep -Evq '^\s*#|^\s*$' $PATH_FILE_ETC_EXPORTS
-  exit_on_failure "$PATH_FILE_ETC_EXPORTS has no exports"
+  on_failure bail "$PATH_FILE_ETC_EXPORTS has no exports"
 
   # ensure we have CAP_SYS_ADMIN
   capsh --print | grep -Eq "^Current: = .*,?cap_sys_admin(,|$)"
-  exit_on_failure 'missing CAP_SYS_ADMIN. be sure to run Docker with --cap-add SYS_ADMIN or --privileged'
+  on_failure bail 'missing CAP_SYS_ADMIN. be sure to run this image with --cap-add SYS_ADMIN or --privileged'
 
   # perform Kerberos assertions
   assert_kerberos_requirements
@@ -395,17 +419,15 @@ boot_helper_mount() {
   local -r type=$(basename "$path")
   local -r args=('-vt' "$type" "$path")
 
-  log "mounting $type onto $path"
+  log "mounting $type filesystem onto $path"
   mount "${args[@]}"
-  stop_on_failure "unable to mount $type onto $path"
+  on_failure stop "unable to mount $type filesystem onto $path"
 }
 
 boot_helper_get_version_flags() {
 
-  local versionFlags
   local -r requestedVersion="$(get_reqd_nfs_version)"
-
-  versionFlags=('--nfs-version' "$requestedVersion" '--no-nfs-version' 2)
+  local versionFlags=('--nfs-version' "$requestedVersion" '--no-nfs-version' 2)
 
   if [[ -z "$(is_nfs3_enabled)" ]]; then
     versionFlags+=('--no-nfs-version' 3)
@@ -432,9 +454,9 @@ boot_main_mounts() {
 
 boot_main_exportfs() {
 
-  log 'exporting filesystems'
+  log 'exporting filesystem(s)'
   $PATH_BIN_EXPORTFS -arv
-  stop_on_failure 'exportfs failed'
+  on_failure stop 'exportfs failed'
 }
 
 boot_main_mountd() {
@@ -442,13 +464,12 @@ boot_main_mountd() {
   local versionFlags
   read -r -a versionFlags <<< "$(boot_helper_get_version_flags)"
   local -r port=$(get_reqd_mountd_port)
-  local -r version=$(get_reqd_nfs_version)
   local -r args=('--debug' 'all' '--port' "$port" "${versionFlags[@]}")
 
   # yes, rpc.mountd is required even for NFS v4: https://forums.gentoo.org/viewtopic-p-7724856.html#7724856
-  log "starting rpc.mountd for NFS version $version on port $port"
+  log "starting rpc.mountd on port $port"
   $PATH_BIN_MOUNTD "${args[@]}"
-  stop_on_failure 'rpc.mountd failed'
+  on_failure stop 'rpc.mountd failed'
 }
 
 boot_main_rpcbind() {
@@ -458,7 +479,7 @@ boot_main_rpcbind() {
   # so as a workaround we start rpcbind now and (in v4-only scenarios) kill it after nfsd starts up
   log 'starting rpcbind'
   $PATH_BIN_RPCBIND -ds
-  stop_on_failure 'rpcbind failed'
+  on_failure stop 'rpcbind failed'
 }
 
 boot_main_idmapd() {
@@ -466,7 +487,7 @@ boot_main_idmapd() {
   if [[ "$(get_reqd_nfs_version)" != '3' && -f "$PATH_FILE_ETC_IDMAPD_CONF" ]]; then
     log 'starting idmapd'
     $PATH_BIN_IDMAPD -v -S
-    stop_on_failure 'idmapd failed'
+    on_failure stop 'idmapd failed'
   fi
 }
 
@@ -480,9 +501,9 @@ boot_main_statd() {
   local -r outPort=$(get_reqd_statd_out_port)
   local -r args=('--no-notify' '--port' "$inPort" '--outgoing-port' "$outPort")
 
-  log "starting statd on port $inPort (outgoing connections on port $outPort)"
+  log "starting statd on port $inPort (outgoing connections from port $outPort)"
   $PATH_BIN_STATD "${args[@]}"
-  stop_on_failure 'statd failed'
+  on_failure stop 'statd failed'
 }
 
 boot_main_nfsd() {
@@ -491,12 +512,11 @@ boot_main_nfsd() {
   read -r -a versionFlags <<< "$(boot_helper_get_version_flags)"
   local -r threads=$(get_reqd_nfsd_threads)
   local -r port=$(get_reqd_nfsd_port)
-  local -r version=$(get_reqd_nfs_version)
   local -r args=('--debug' 8 '--port' "$port" "${versionFlags[@]}" "$threads")
 
-  log "starting rpc.nfsd on port $port with version $version and $threads server thread(s)"
+  log "starting rpc.nfsd on port $port with $threads server thread(s)"
   $PATH_BIN_NFSD "${args[@]}"
-  stop_on_failure 'rpc.nfsd failed'
+  on_failure stop 'rpc.nfsd failed'
 
   if [ -z "$(is_nfs3_enabled)" ]; then
     kill_process_if_running "$PATH_BIN_RPCBIND"
@@ -511,14 +531,63 @@ boot_main_svcgssd() {
 
   log 'starting rpc.svcgssd'
   $PATH_BIN_RPC_SVCGSSD -f &
-  stop_on_failure 'rpc.svcgssd failed'
+  on_failure stop 'rpc.svcgssd failed'
 }
 
-boot_main_print_ready_message() {
 
-  log_header "ready and waiting for connections on port $(get_reqd_nfsd_port)"
-  log 'list of exports:'
-  cat $PATH_FILE_ETC_EXPORTS
+######################################################################################
+### boot summary
+######################################################################################
+
+summarize_nfs_versions() {
+
+  local -r reqd_version="$(get_reqd_nfs_version)"
+  local versions=''
+
+  case "$reqd_version" in
+    4\.2)
+      versions='4.2, 4.1, 4'
+      ;;
+    4\.1)
+      versions='4.1, 4'
+      ;;
+    4)
+      versions='4'
+      ;;
+    *)
+      versions='3'
+      ;;
+  esac
+
+  if [[ -n "$(is_nfs3_enabled)" && "$reqd_version" =~ ^4 ]]; then
+    versions="$versions, 3"
+  fi
+
+  log "list of enabled NFS protocol versions: $versions"
+}
+
+summarize_exports() {
+
+  log 'list of container exports:'
+
+  while read -r export; do
+    log "  $export"
+  done < "$PATH_FILE_ETC_EXPORTS"
+}
+
+summarize_ports() {
+
+  local -r nfsdPort="$(get_reqd_nfsd_port)"
+
+  if [[ -z "$(is_nfs3_enabled)" ]]; then
+    log "list of container ports that should be exposed: $nfsdPort (TCP)"
+  else
+    log 'list of container ports that should be exposed:'
+    log '  111 (TCP and UDP)'
+    log "  $nfsdPort (TCP and UDP)"
+    log "  $(get_reqd_statd_in_port) (TCP and UDP)"
+    log "  $(get_reqd_mountd_port) (TCP and UDP)"
+  fi
 }
 
 
@@ -549,10 +618,22 @@ boot() {
   boot_main_idmapd
   boot_main_nfsd
   boot_main_svcgssd
-  boot_main_print_ready_message
+
+  log 'all services started normally'
+}
+
+summarize() {
+
+  log_header 'server startup complete'
+
+  summarize_nfs_versions
+  summarize_exports
+  summarize_ports
 }
 
 hangout() {
+
+  log_header 'ready and waiting for NFS client connections'
 
   # wait forever or until we get SIGTERM or SIGINT
   # https://stackoverflow.com/a/41655546/229920
@@ -564,6 +645,7 @@ main() {
 
   init
   boot
+  summarize
   hangout
 }