+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+* upon successful server startup, log NFS versions, exports, and ports
+
+### Fixed
+### Changed
+
 ## [1.1.1] - 2018-08-21
 
 ### Fixed